moving media layer out of transport namespace
